Refactor a big chunk of the dellnos9 and brocade driver.
Move a lot of the common stuff to _unconsole.session (name will probably
change).

This disables saving-to-switch in the dellnos9 but doesn't remove all the
related methods yet.
